welcome/login/register code is now modular
